requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
dj-database-url==0.5.0
Django==2.2.1
django-ckeditor==6.0.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.9.4
gunicorn==20.0.4
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
whitenoise==5.2.0
wrapt==1.12.1
